refactor: move error boundaries to features dir
