+++ packages/slice-machine/src/features/errorBoundaries/AppStateErrorBoundary.tsx
@@ -9,54 +9,12 @@ import {
 } from "@prismicio/editor-ui";
 import {
   isInvalidActiveEnvironmentError,
-  isUnauthenticatedError,
   isUnauthorizedError,
 } from "@slicemachine/manager/client";
 import Link from "next/link";
-import {
-  type ComponentPropsWithoutRef,
-  type FC,
-  PropsWithChildren,
-  useCallback,
-  useRef,
-} from "react";
+import { PropsWithChildren } from "react";
 
 import { ReloadLogoutButton } from "@/features/auth/LogoutButton";
-import { useAuthStatus } from "@/hooks/useAuthStatus";
-
-type DefaultErrorBoundaryProps = Pick<
-  // TODO(DT-1979): Export the `ErrorBoundaryProps` type from `@prismicio/editor-ui`.
-  ComponentPropsWithoutRef<typeof EditorUiErrorBoundary>,
-  "children" | "renderError"
->;
-
-export const DefaultErrorBoundary: FC<DefaultErrorBoundaryProps> = (props) => {
-  const errorRef = useRef<unknown>();
-  const authStatus = useAuthStatus();
-  return (
-    <EditorUiErrorBoundary
-      {...props}
-      onError={(error) => {
-        errorRef.current = error;
-      }}
-      ref={useCallback(
-        (errorBoundary: EditorUiErrorBoundary | null) => {
-          if (errorBoundary === null) return;
-          const error = errorRef.current;
-          if (isUnauthenticatedError(error) || isUnauthorizedError(error)) {
-            errorRef.current = undefined;
-            errorBoundary.reset();
-          }
-        },
-        /* eslint-disable-next-line react-hooks/exhaustive-deps --
-         * Whenever `authStatus` changes, we want to `reset` the `errorBoundary`
-         * if an authentication or authorization error has been caught.
-         **/
-        [authStatus],
-      )}
-    />
-  );
-};
 
 export function AppStateErrorBoundary(props: PropsWithChildren) {
   return (
