+++ packages/slice-machine/src/features/errorBoundaries/index.ts
@@ -1,0 +1,2 @@
+export { AppStateErrorBoundary } from "./AppStateErrorBoundary";
+export { DefaultErrorBoundary } from "./DefaultErrorBoundary";
